pyproject.toml
[tool.poetry]
name = "create-readme"
version = "0.1.0"
description = ""
authors = ["rizerphe <44440399+rizerphe@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "create_readme"}]

[tool.poetry.dependencies]
python = "^3.11"
click-config-file = "^0.6.0"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create-readme = "create_readme.run:run"
